+++ routers/post.py
@@ -34,10 +34,14 @@ def get_post(post_id):
 
 
 @router.post("/add")
-def add_post(post: Post):
+def add_post(post: Post, request: Request):
     try:
-        doc_ref = db.collection(u"posts")
-        doc_ref.add(dict(post))
+        uid = request.headers.get("uid")
+        doc = db.collection(u"users").document(uid).get().to_dict()
+        if doc["admin"] == True:
+            doc_ref = db.collection(u"posts")
+            doc_ref.add(dict(post))
+        raise Exception()
 
     except Exception as e:
         print(e)
@@ -45,17 +49,21 @@ def add_post(post: Post):
 
 
 @router.put("/{post_id}")
-def edit_post(post_id, post: Post):
+def edit_post(post_id, post: Post, request: Request):
     try:
-        edit = db.collection(u"posts").document(post_id)
-        new_data = post.dict(exclude_none=True, exclude_defaults=True)
+        uid = request.headers.get("uid")
+        doc = db.collection(u"users").document(uid).get().to_dict()
+        if doc["admin"] == True:
+            edit = db.collection(u"posts").document(post_id)
+            new_data = post.dict(exclude_none=True, exclude_defaults=True)
 
-        if "created_at" in new_data or "updated_at" in new_data:
-            raise Exception()
+            if "created_at" in new_data or "updated_at" in new_data:
+                raise Exception()
 
-        new_data["updated_at"] = datetime.now(timezone("Asia/Kolkata"))
+            new_data["updated_at"] = datetime.now(timezone("Asia/Kolkata"))
 
-        edit.update(dict(new_data))
+            edit.update(dict(new_data))
+        raise Exception()
 
     except Exception as e:
         print(e)
@@ -63,9 +71,14 @@ def edit_post(post_id, post: Post):
 
 
 @router.delete("/{post_id}")
-def delete_post(post_id):
+def delete_post(post_id, request: Request):
     try:
-        db.collection(u"posts").document(post_id).delete()
+        uid = request.headers.get("uid")
+        doc = db.collection(u"users").document(uid).get().to_dict()
+        if doc["admin"] == True:
+            db.collection(u"posts").document(post_id).delete()
+        else:
+            raise Exception()
 
     except Exception as e:
         print(e)
